issue #96 rename : placeInsert -> AddPlace
