+++ client/app/src/main/res/layout/fragment_add_place.xml
@@ -6,7 +6,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.ui.search.PlaceInsertFragment">
+        tools:context=".presentation.ui.search.AddPlaceFragment">
 
         <androidx.constraintlayout.widget.ConstraintLayout
             android:id="@+id/mapView"
@@ -22,7 +22,7 @@
         </androidx.constraintlayout.widget.ConstraintLayout>
 
         <Button
-            android:id="@+id/button3"
+            android:id="@+id/addPlaceButton"
             android:layout_width="120dp"
             android:layout_height="50dp"
             android:layout_marginTop="20dp"
